Address Argent artifact upload review feedback
