+++ packages/build-tools/src/steps/utils/remoteDeviceRunSession.ts
@@ -8,9 +8,6 @@ import nullthrows from 'nullthrows';
 import { z } from 'zod';
 import { randomBytes } from 'node:crypto';
 import fs from 'node:fs';
-import os from 'node:os';
-import path from 'node:path';
-import fetch, { Headers } from 'node-fetch';
 
 import { CustomBuildContext } from '../../customBuildContext';
 import { Sentry } from '../../sentry';
@@ -30,48 +27,6 @@ const START_DEVICE_RUN_SESSION_MUTATION = graphql(`
     }
   }
 `);
-
-const CREATE_DEVICE_RUN_SESSION_ARTIFACT_UPLOAD_SESSION_MUTATION = graphql(`
-  mutation CreateDeviceRunSessionArtifactUploadSession(
-    $deviceRunSessionId: ID!
-    $input: CreateDeviceRunSessionArtifactUploadSessionInput!
-  ) {
-    deviceRunSession {
-      createArtifactUploadSession(deviceRunSessionId: $deviceRunSessionId, input: $input) {
-        uploadSession {
-          url
-          headers
-        }
-      }
-    }
-  }
-`);
-
-const ARGENT_FLAGS_FILE = path.join(os.homedir(), '.argent', 'flags.json');
-const ARGENT_ARTIFACTS_LIST_ENDPOINT_FLAG = 'artifacts-list-endpoint';
-const ARGENT_ARTIFACT_UPLOAD_POLL_INTERVAL_MS = 5_000;
-
-const ArgentArtifactSchema = z.object({
-  id: z.string(),
-  filename: z.string(),
-  mimeType: z.string(),
-  size: z.number(),
-  isDirectory: z.boolean().optional(),
-});
-const ArgentArtifactsListResponseSchema = z.object({
-  artifacts: z.array(ArgentArtifactSchema),
-});
-const ArtifactUploadSessionSchema = z.object({
-  url: z.string(),
-  headers: z.record(z.string(), z.string()),
-});
-
-type ArgentArtifact = z.infer<typeof ArgentArtifactSchema>;
-
-type ArtifactUploadSession = {
-  url: string;
-  headers: Record<string, string>;
-};
 
 export function getDeviceRunSessionIdOrThrow(env: BuildStepEnv): string {
   const deviceRunSessionId = env.DEVICE_RUN_SESSION_ID;
@@ -107,37 +62,6 @@ export function getNgrokAuthtokenOrThrow(env: BuildStepEnv): string {
     );
   }
   return authtoken;
-}
-
-export async function enableArgentArtifactsListEndpointAsync(): Promise<void> {
-  let flags: Record<string, boolean> = {};
-  try {
-    const raw = await fs.promises.readFile(ARGENT_FLAGS_FILE, 'utf8');
-    const parsed = JSON.parse(raw) as unknown;
-    if (
-      parsed &&
-      typeof parsed === 'object' &&
-      !Array.isArray(parsed) &&
-      'flags' in parsed &&
-      parsed.flags &&
-      typeof parsed.flags === 'object' &&
-      !Array.isArray(parsed.flags)
-    ) {
-      flags = Object.fromEntries(
-        Object.entries(parsed.flags).filter((entry): entry is [string, boolean] => {
-          return typeof entry[1] === 'boolean';
-        })
-      );
-    }
-  } catch (err) {
-    if (!(err instanceof Error && 'code' in err && err.code === 'ENOENT')) {
-      throw err;
-    }
-  }
-
-  flags[ARGENT_ARTIFACTS_LIST_ENDPOINT_FLAG] = true;
-  await fs.promises.mkdir(path.dirname(ARGENT_FLAGS_FILE), { recursive: true });
-  await fs.promises.writeFile(ARGENT_FLAGS_FILE, `${JSON.stringify({ flags }, null, 2)}\n`, 'utf8');
 }
 
 const TurnIceServersSchema = z.array(
@@ -237,234 +161,6 @@ export async function fetchServeSimTurnArgsAsync(
     );
     return [];
   }
-}
-
-export function startArgentArtifactUploadPollingAsync({
-  ctx,
-  deviceRunSessionId,
-  toolsUrl,
-  toolsAuthToken,
-  logger,
-}: {
-  ctx: CustomBuildContext;
-  deviceRunSessionId: string;
-  toolsUrl: string;
-  toolsAuthToken?: string;
-  logger: bunyan;
-}): void {
-  const uploadedArtifactIds = new Set<string>();
-  const uploadingArtifactIds = new Set<string>();
-
-  void pollArgentArtifactsForUploadAsync({
-    ctx,
-    deviceRunSessionId,
-    toolsUrl,
-    toolsAuthToken,
-    uploadedArtifactIds,
-    uploadingArtifactIds,
-    logger,
-  });
-}
-
-export async function pollArgentArtifactsForUploadAsync({
-  ctx,
-  deviceRunSessionId,
-  toolsUrl,
-  toolsAuthToken,
-  uploadedArtifactIds,
-  uploadingArtifactIds,
-  logger,
-}: {
-  ctx: CustomBuildContext;
-  deviceRunSessionId: string;
-  toolsUrl: string;
-  toolsAuthToken?: string;
-  uploadedArtifactIds: Set<string>;
-  uploadingArtifactIds: Set<string>;
-  logger: bunyan;
-}): Promise<never> {
-  logger.info('Polling Argent tool-server for artifacts every 5 seconds.');
-
-  for (;;) {
-    try {
-      const artifacts = await listArgentArtifactsAsync({ toolsUrl, toolsAuthToken });
-      await Promise.all(
-        artifacts
-          .filter(artifact => !uploadedArtifactIds.has(artifact.id))
-          .filter(artifact => !uploadingArtifactIds.has(artifact.id))
-          .map(async artifact => {
-            uploadingArtifactIds.add(artifact.id);
-            try {
-              await uploadArgentArtifactAsync({
-                ctx,
-                deviceRunSessionId,
-                toolsUrl,
-                toolsAuthToken,
-                artifact,
-                logger,
-              });
-              uploadedArtifactIds.add(artifact.id);
-            } finally {
-              uploadingArtifactIds.delete(artifact.id);
-            }
-          })
-      );
-    } catch (err) {
-      const error = err instanceof Error ? err : new Error(String(err));
-      Sentry.capture('Could not upload Argent remote session artifacts', error, {
-        level: 'warning',
-      });
-      logger.warn({ err: error }, 'Could not upload Argent remote session artifacts.');
-    }
-    await sleepAsync(ARGENT_ARTIFACT_UPLOAD_POLL_INTERVAL_MS);
-  }
-}
-
-export async function listArgentArtifactsAsync({
-  toolsUrl,
-  toolsAuthToken,
-}: {
-  toolsUrl: string;
-  toolsAuthToken?: string;
-}): Promise<ArgentArtifact[]> {
-  const response = await fetch(new URL('/artifacts', toolsUrl).toString(), {
-    headers: getArgentToolServerAuthHeaders(toolsAuthToken),
-  });
-  if (!response.ok) {
-    throw new Error(`Failed to list Argent artifacts: ${response.status} ${response.statusText}`);
-  }
-  const result = ArgentArtifactsListResponseSchema.safeParse(await response.json());
-  if (!result.success) {
-    throw new Error(`Invalid Argent artifacts response: ${result.error.message}`);
-  }
-  return result.data.artifacts;
-}
-
-export async function uploadArgentArtifactAsync({
-  ctx,
-  deviceRunSessionId,
-  toolsUrl,
-  toolsAuthToken,
-  artifact,
-  logger,
-}: {
-  ctx: CustomBuildContext;
-  deviceRunSessionId: string;
-  toolsUrl: string;
-  toolsAuthToken?: string;
-  artifact: ArgentArtifact;
-  logger: bunyan;
-}): Promise<void> {
-  const filename = getArgentArtifactUploadFilename(artifact);
-  logger.info(`Uploading Argent artifact ${filename} (${artifact.size} bytes).`);
-  const stream = await createArgentArtifactDownloadStreamAsync({
-    artifact,
-    toolsUrl,
-    toolsAuthToken,
-  });
-  const uploadSession = await createDeviceRunSessionArtifactUploadSessionAsync({
-    ctx,
-    deviceRunSessionId,
-    artifactId: artifact.id,
-    filename,
-    size: artifact.size,
-  });
-  await uploadArtifactToSignedUrlAsync({
-    uploadSession,
-    stream,
-  });
-}
-
-async function createArgentArtifactDownloadStreamAsync({
-  artifact,
-  toolsUrl,
-  toolsAuthToken,
-}: {
-  artifact: ArgentArtifact;
-  toolsUrl: string;
-  toolsAuthToken?: string;
-}): Promise<NodeJS.ReadableStream> {
-  const response = await fetch(new URL(`/artifacts/${artifact.id}`, toolsUrl).toString(), {
-    headers: getArgentToolServerAuthHeaders(toolsAuthToken),
-  });
-  if (!response.ok) {
-    throw new Error(
-      `Failed to download Argent artifact ${artifact.id}: ${response.status} ${response.statusText}`
-    );
-  }
-  if (!response.body) {
-    throw new Error(`Argent artifact ${artifact.id} response did not include a readable body.`);
-  }
-  return response.body;
-}
-
-function getArgentArtifactUploadFilename(artifact: ArgentArtifact): string {
-  return artifact.isDirectory ? `${artifact.filename}.tar.gz` : artifact.filename;
-}
-
-async function createDeviceRunSessionArtifactUploadSessionAsync({
-  ctx,
-  deviceRunSessionId,
-  artifactId,
-  filename,
-  size,
-}: {
-  ctx: CustomBuildContext;
-  deviceRunSessionId: string;
-  artifactId: string;
-  filename: string;
-  size: number;
-}): Promise<ArtifactUploadSession> {
-  const result = await ctx.graphqlClient
-    .mutation(CREATE_DEVICE_RUN_SESSION_ARTIFACT_UPLOAD_SESSION_MUTATION, {
-      deviceRunSessionId,
-      input: {
-        name: `Argent artifact ${filename} (${artifactId})`,
-        filename,
-        size,
-      },
-    })
-    .toPromise();
-  if (result.error) {
-    throw new SystemError(
-      `Failed to create upload session for Argent artifact ${artifactId}: ${result.error.message}`
-    );
-  }
-  const uploadSession =
-    result.data?.deviceRunSession.createArtifactUploadSession.uploadSession ?? null;
-  if (!uploadSession) {
-    throw new SystemError(`Upload session for Argent artifact ${artifactId} was not returned.`);
-  }
-  const parsedUploadSession = ArtifactUploadSessionSchema.safeParse(uploadSession);
-  if (!parsedUploadSession.success) {
-    throw new SystemError(
-      `Upload session for Argent artifact ${artifactId} was invalid: ${parsedUploadSession.error.message}`
-    );
-  }
-  return parsedUploadSession.data;
-}
-
-async function uploadArtifactToSignedUrlAsync({
-  uploadSession,
-  stream,
-}: {
-  uploadSession: ArtifactUploadSession;
-  stream: NodeJS.ReadableStream;
-}): Promise<void> {
-  const response = await fetch(uploadSession.url, {
-    method: 'PUT',
-    headers: new Headers(uploadSession.headers),
-    body: stream,
-  });
-  if (!response.ok) {
-    throw new Error(`Failed to upload Argent artifact: ${response.status} ${response.statusText}`);
-  }
-}
-
-function getArgentToolServerAuthHeaders(
-  toolsAuthToken: string | undefined
-): Record<string, string> {
-  return toolsAuthToken ? { Authorization: `Bearer ${toolsAuthToken}` } : {};
 }
 
 export async function uploadRemoteSessionConfigAsync({
